:sparkles: :hammer: Creating command line options
Creating manager.js which will be the only executable file which
receives command line options to know which of the functionalities it
should execute. For now there are total user balance, price and
percentage and bitcoin wallet manager.
Besides that, excluded some useless functions which were leftovers from
last refactoring.